boldi

+++ src/main/java/com/example/student_management/repository/StudentRepo.java
@@ -3,5 +3,6 @@ package com.example.student_management.repository;
 import com.example.student_management.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepo extends JpaRepository<Student, Integer> {
-}
+public interface StudentRepo extends JpaRepository<Student, Long> {
+    }
+
